api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.5",
    "google-genai>=1.62.0",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-openai>=0.3",
    "pillow>=12.1.0",
    "pydantic>=2.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
